requirements.txt
python-dotenv==0.10.3
inflection==0.3.1
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
kubernetes==10.0.0
MarkupSafe==1.1.1
oauthlib==3.0.2
pyasn1==0.4.5
pyasn1-modules==0.2.5
python-dateutil==2.8.0
PyYAML==5.1.1
rsa==4.0
six==1.12.0
urllib3==1.25.3
websocket-client==0.56.0
Werkzeug==0.15.5
python-engineio==3.4.3